# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rest API",
        "contact": {
            "name": "Md Amjad hossain",
            "url": ""
        },
        "version": "1.0"
    },
    "host": "localhost:8080/",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "api/injuryactivity": {
            "get": {
                "tags": [
                    "InjuryActivity"
                ],
                "summary": "Find All InjuryActivity",
                "description": "Returns InjuryActivity overview.",
                "operationId": "index",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "InjuryActivity overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "InjuryActivity"
                ],
                "summary": "Update an existing InjuryActivity",
                "description": "",
                "operationId": "updateInjuryActivity",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "household_unique_code is mandatory otherwise you will loose data object that needs to be added to the store[f01-f33]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InjuryMortality"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "InjuryMortality not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "api/injuryactivity/{household_unique_code}": {
            "get": {
                "tags": [
                    "InjuryActivity"
                ],
                "summary": "Find All InjuryActivity",
                "description": "Returns InjuryActivity overview.",
                "operationId": "getInjuryActivities",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "InjuryActivity overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "api/injuryactivity/roadtransportinjury/{household_unique_code}": {
            "put": {
                "tags": [
                    "InjuryActivity"
                ],
                "summary": "Update an existing InjuryActivity",
                "description": "",
                "operationId": "updateInjuryActivity",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "InjuryActivity object that needs to be added to the store[h01-h09]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InjuryMortality"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "InjuryMortality not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    },
...
}